Hometask_7/

+++ Hometask_7/Hometask_7.py
@@ -3,6 +3,7 @@ import pathlib
 import os
 from collections import Counter
 import pandas as pd
+from decimal import *
 
 
 PATH = os.path.join(pathlib.Path.cwd(), 'newsfeed.txt')
@@ -60,7 +61,7 @@ def percent_of_total_cnt(dictionary):
     new_dict = {}
     total_cnt = sum(dictionary.values())
     for letter, count in dictionary.items():
-        percent = round(count * 100.0 / total_cnt)
+        percent = round(Decimal(count * 100.0 / total_cnt))
         new_dict[letter] = percent
 
     return new_dict
